Rename CVRPLIB to VRPLIB
