+++ .github/workflows/vrplib.yaml
@@ -1,4 +1,4 @@
-name: CVRPLIB
+name: VRPLIB
 
 on:
   push:
