+++ vrplib/download/download_instance.py
@@ -1,8 +1,8 @@
 from functools import lru_cache
 from urllib.request import urlopen
 
-from cvrplib.constants import CVRPLIB_URL
-from cvrplib.parse import parse_solomon, parse_vrplib
+from vrplib.constants import CVRPLIB_URL
+from vrplib.parse import parse_solomon, parse_vrplib
 
 from .download_utils import find_set, is_vrptw
 
@@ -15,7 +15,7 @@ def download_instance(name: str):
     Parameters
     ----------
     name
-        The instance name. See `cvrplib.list_instances` for all eligible names.
+        The instance name. See `vrplib.list_instances` for all eligible names.
 
     Returns
     -------
