+++ vrplib/download/download_solution.py
@@ -1,8 +1,8 @@
 from functools import lru_cache
 from urllib.request import urlopen
 
-from cvrplib.constants import CVRPLIB_URL
-from cvrplib.parse import parse_solution
+from vrplib.constants import CVRPLIB_URL
+from vrplib.parse import parse_solution
 
 from .download_utils import find_set
 
@@ -15,7 +15,7 @@ def download_solution(name: str):
     Parameters
     ----------
     name
-        The instance name. See `cvrplib.list_instances` for all eligible names.
+        The instance name. See `vrplib.list_instances` for all eligible names.
 
     Returns
     -------
